requirements.txt
flask==2.3.3
flask-cors==4.0.0
pymongo==4.5.0
pymysql==1.1.0
kafka-python==2.0.2
pyspark==3.4.1
pandas==2.0.3
numpy==1.24.3
plotly==5.15.0
python-dotenv==1.0.0
bcrypt==4.0.1